move assertion to post deployment
